Removed unnecessary PORTALSinteraction scripts
